+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- [Add command for revealing the **Calva says** output channel](https://github.com/BetterThanTomorrow/calva/issues/2436)
+
 ## [2.0.423] - 2024-03-17
 
 - [Add Output Destinations configuration](https://github.com/BetterThanTomorrow/calva/issues/1104)